+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 - Fixed KiCad calibration points would be reset after using component label editor
 - Changed banner for updating data from online source to show specific files being updated
+- Changed "Clear All" button will not be shown in collapsed "Important signals" when no selected signals